Rename pipeline.Resource to ingest.Pipeline
