+++ internal/elasticsearch/ingest/pipeline.go
@@ -2,7 +2,7 @@
 // or more contributor license agreements. Licensed under the Elastic License;
 // you may not use this file except in compliance with the Elastic License.
 
-package pipeline
+package ingest
 
 import (
 	"encoding/json"
@@ -12,15 +12,15 @@ import (
 	"gopkg.in/yaml.v3"
 )
 
-// Resource represents a pipeline resource loaded from a file
-type Resource struct {
+// Pipeline represents a pipeline resource loaded from a file
+type Pipeline struct {
 	Name    string // Name of the pipeline
 	Format  string // Format (extension) of the pipeline
 	Content []byte // Content is the original file contents.
 }
 
 // Filename returns the original filename associated with the pipeline.
-func (p *Resource) Filename() string {
+func (p *Pipeline) Filename() string {
 	pos := strings.LastIndexByte(p.Name, '-')
 	if pos == -1 {
 		pos = len(p.Name)
@@ -29,7 +29,7 @@ func (p *Resource) Filename() string {
 }
 
 // MarshalJSON returns the pipeline contents in JSON format.
-func (p *Resource) MarshalJSON() (asJSON []byte, err error) {
+func (p *Pipeline) MarshalJSON() (asJSON []byte, err error) {
 	switch p.Format {
 	case "json":
 		asJSON = p.Content
