+++ internal/testrunner/runners/pipeline/runner.go
@@ -7,6 +7,7 @@ package pipeline
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/elastic/elastic-package/internal/elasticsearch/ingest"
 	"os"
 	"path/filepath"
 	"regexp"
@@ -16,7 +17,6 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/elastic/elastic-package/internal/common"
-	"github.com/elastic/elastic-package/internal/elasticsearch/pipeline"
 	"github.com/elastic/elastic-package/internal/fields"
 	"github.com/elastic/elastic-package/internal/logger"
 	"github.com/elastic/elastic-package/internal/multierror"
@@ -31,7 +31,7 @@ const (
 
 type runner struct {
 	options   testrunner.TestOptions
-	pipelines []pipeline.Resource
+	pipelines []ingest.Pipeline
 }
 
 func (r *runner) TestFolderRequired() bool {
